Making logbook more tolerant to failures
The architecture that logbook is based on is 'fail early'. That is,
happily crash the application if something fails. Clients are expected
to reconnect gracefully.

To assert that logbook gracefully goes down on errors, the `run(...)`
function is now elegantly wrapped in `with` statement(s) that makes sure
that initialization and, more so, cleanup of ZeroMQ context and sockets
is done properly. Without this, things would not have been correctly
cleaned up.

This commit also refactored most of the `run(..)` method into a helper
class called `_LogBookRunner`. Doing this made it easier to 1) understand
which state is necessary for the application to actually run and 2) more
easily split the runnable code into smaller logical managable chunks.

Also changed the context to run three threads instead of one. Since
three sockets are involved in shuffling data, this felt more reasonable.
Whether the change has any bigger effect could be meassured in the
future.