Created `RotatedEventStore`
`RotatedEventStore` was factored out of the, still existing,
`PersistedEventStore`. `PersistedEventStore` now simply takes care of
synchronizing rotation of the `RotatedEventStore`s it wraps.

`PersistedEventStore` does rotation based on the number of events that
has previously been written to an event store. This number is stored in
an internal counter.

All tests are OK.