Make sure all Py3-friendly code runs on Py2
This is checked by running all tests in Python 2.7.

The issue was that `zmq_socket.recv()` returns a `str` instance in
Python 2 and a `bytes` instance in Python 3. Therefor, I had to have an
if-statement taking care converting to `str` when needed. It is worth
pointing out that this is the first `if`-statement I've had to write
when making Rewind Python 3 compatible!