Events are added through query (RES) socket
This was done after thorough consideration on how to transmit new
messages into Rewind. Two key things played a part in this:

 * A clean shutdown would be impossible with PUSH/PULL sockets for event
   transmission. See
   http://stackoverflow.com/questions/13229918/shut-down-zeromq-receiving-end-without-loss
   for details.
 * Also, each event is so precious that is should probably be
   acknowledged that is arrived at the Rewind end correctly. A two phase
   commit for new events might a possible solution in the future to
   handle split brain syndrome if a client crashes after publishing an
   event (or before it's event has been sent).

The flipside of using REQ/RES sockets for event transmission is that it
makes writing events synchronous instead of async. However, there is
nothing that says each client can't kickstart an internal PULL-thread
for queuing outgoing events.

It should also be noted that using a ZMQ_QUEUE device would make is
possible to get low latencies when publishing events.